working on writing on files

+++ src/io/CsvHandler.java
@@ -1,5 +1,6 @@
 package io;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 
@@ -8,26 +9,32 @@ public class CsvHandler {
     FileWriter csvFile;
     String fileName;
     String filePath;
-    CsvHandler(String fileName) {
+
+    File file;
+
+    public CsvHandler(String fileName) {
         this.fileName = fileName;
         this.createFile(fileName);
     }
 
-    private void createFile(String fileName){
-        try{
-            this.csvFile = new FileWriter(fileName);
+    private void createFile(String fileName) {
+        try {
+            this.file = new File(fileName);
+            if (!this.file.isFile()) {
+                this.csvFile = new FileWriter(fileName);
+            }
 
-        } catch(IOException e){
+        } catch (IOException e) {
             System.out.println("Error occured");
             e.printStackTrace();
         }
     }
 
-    public void writeToFile(String data){
-        try{
+    public void writeToFile(String data) {
+        try {
             this.csvFile.write(data);
             this.csvFile.close();
-        } catch(IOException e){
+        } catch (IOException e) {
             System.out.println("Error occured");
             e.printStackTrace();
         }
